requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
pydantic>=2
python-multipart
Jinja2
starlette
openpyxl
python-docx
xhtml2pdf
